refactor: TIO-264: split dialog helpers out of util_functions
util_functions.dart had become a grab-bag — number formatting,
navigation, date helpers, an icon factory, and five unrelated
showDialog wrappers all in one file. Group the dialogs into their own
module so callers express intent ("I want a dialog helper") through
the import name, and so future dialog additions have an obvious home
instead of growing the catch-all further.

Move editTwoTitles, _submitTitles, askForSavingQuickTool,
showFileNotAccessibleDialog, and showNoCameraFoundDialog into
lib/util/dialogs.dart. util_functions.dart retains formatSettingValues,
openSettingPage, getCurrentDateTime, circleToolIcon, IncreaseOrDecrease,
and blockValuesSameAsDefaultBlock — all non-dialog. Drops it from 237
to 101 lines (under the 200-line limit). The four caller pages
(piano, parent_tool, image_page, media_player_page) gain a direct
import of dialogs.dart.

MAX_FILE_COUNT 23 -> 22. MAX_AVG_LENGTH 345 -> 350 (mechanically higher
because the removed file was below the previous average; the per-file
200-line limit is what we actually enforce).
