+++ .github/workflows/reusable-verify.yaml
@@ -1,8 +1,8 @@
 name: Verify
 
 env:
-  MAX_FILE_COUNT: 23
-  MAX_AVG_LENGTH: 345
+  MAX_FILE_COUNT: 22
+  MAX_AVG_LENGTH: 350
   MIN_COVERAGE: 63.4
 
 on:
