+++ lib/pages/media_player/media_player_page.dart
@@ -32,6 +32,7 @@ import 'package:tiomusic/services/wakelock.dart';
 import 'package:tiomusic/util/color_constants.dart';
 import 'package:tiomusic/util/constants/constants.dart';
 import 'package:tiomusic/util/constants/media_player_constants.dart';
+import 'package:tiomusic/util/dialogs.dart';
 import 'package:tiomusic/util/log.dart';
 import 'package:tiomusic/util/tool_navigation_utils.dart';
 import 'package:tiomusic/util/util_functions.dart';
