+++ lib/pages/parent_tool/parent_tool.dart
@@ -14,6 +14,7 @@ import 'package:tiomusic/services/file_system.dart';
 import 'package:tiomusic/services/project_repository.dart';
 import 'package:tiomusic/util/color_constants.dart';
 import 'package:tiomusic/util/constants/constants.dart';
+import 'package:tiomusic/util/dialogs.dart';
 import 'package:tiomusic/util/util_functions.dart';
 import 'package:tiomusic/util/tutorial/tutorial_util.dart';
 import 'package:tiomusic/widgets/card_list_tile.dart';
